Chemistry

Constants

Polymath.Chemistry.AVOGADRO — Constant

Constant representing Avogadro's number.

Example

julia> AVOGADRO
6.02214076e23

Functions

Polymath.Chemistry.atoms_to_moles — Function

atoms_to_moles(atoms)

Calculate number of moles given the quantity of atoms or molecules.

Example

julia> 3.613284456e24 / AVOGADRO
6.0

Polymath.Chemistry.moles_to_atoms — Function

moles_to_atoms(moles)

Calculate number of atoms or molecules given the quantity of moles (mol).

Example

julia> 6 * AVOGADRO
3.613284456e24