Geometry

Usage

To use this module:

include("Geometry.jl")
using .Geometry

Constants

Polymath.Geometry.PI — Constant

Constant representing the value of π (pi).

Example

julia> PI
3.141592653589793

Types

Polymath.Geometry.Point — Type

Point(x, y)
Point(x, y, z)

Structure containing 2D or 3D coordinates.

Examples

julia> p1 = Point(1, 2, 3)
Point(1, 2, 3)

julia> p2 = Point(4, 5)
Point(4, 5, nothing)

Functions

Polymath.Geometry.circumference — Function

circumference(radius)

Calculate circumference given the radius.

Example

julia> circumference(32.50)
204.20352248333654

Polymath.Geometry.diameter — Function

diameter(radius)

Calculate diameter given the radius.

Example

julia> diameter(17.75)
35.5

Polymath.Geometry.radius — Function

radius(diameter)

Calculate radius given the diameter.

Example

julia> radius(54.25)
27.125

Polymath.Geometry.area_circle — Function

area_circle(radius)

Calculate area of a circle given the radius.

Example

julia> area_circle(157)
77437.11731833481

Polymath.Geometry.area_cube — Function

area_cube(side_length)

Calculate area of a cube given the length of a side.

Example

julia> area_cube(42)
10584

Polymath.Geometry.area_ellipse — Function

area_ellipse(major_semiaxis, minor_semiaxis)

Calculate area of an ellipse given the major and minor semiaxes.

Example

julia> area_ellipse(50.575, 50.419)
8010.875677067962

Polymath.Geometry.area_rectangle — Function

area_rectangle(height, width)

Calculate area of a rectangle given the height and width.

Example

julia> area_rectangle(15.75, 20)
315.0

Polymath.Geometry.area_square — Function

area_square(side_length)

Calculate area of a square given the length of a side.

Example

julia> area_square(25.50)
650.25

Polymath.Geometry.area_trapezoid — Function

area_trapezoid(base_length, top_length, height)

Calculate area of a trapezoid given the base length, top length, and height.

Example

julia> area_trapezoid(30.5, 10.5, 20.75)
425.375

Polymath.Geometry.area_triangle — Function

area_triangle(base_length, height)

Calculate area of a triangle given the base length and height.

Example

julia> area_triangle(42.5, 32.50)
690.625

Polymath.Geometry.perimeter_rectangle — Function

perimeter_rectangle(height, width)

Calculate perimeter of a rectangle given the height and width.

Example

julia> perimeter_rectangle(25, 50)
150

Polymath.Geometry.perimeter_square — Function

perimeter_square(side_length)

Calculate perimeter of a square given the length of a side.

Example

julia> perimeter_square(25)
100

Polymath.Geometry.edges — Function

edges(faces, vertices)

Calculate edges of a convex polyhedron given the faces and vertices.

Example

julia> edges(6, 8)
12

Polymath.Geometry.faces — Function

faces(edges, vertices)

Calculate faces of convex polyhedron given the edges and vertices.

Example

julia> faces(12, 8)
6

Polymath.Geometry.vertices — Function

vertices(edges, faces)

Calculate vertices of a convex polyhedron given the edges and faces.

Example

julia> vertices(12, 6)
8

Polymath.Geometry.has_euler_characteristic — Function

has_euler_characteristic(edges, faces, vertices)

Determine whether or not a polyhedron has the Euler characteristic given the edges, faces, and vertices.

Examples

julia> has_euler_characteristic(6, 4, 4) # tetrahedron
true

julia> has_euler_characteristic(12, 7, 6) # tetrahemihexahedron
false

Polymath.Geometry.distance_2d — Function

distance_2d(x1, y1, x2, y2)
distance_2d(p1, p2)

Calculate distance between a pair of 2D points given the coordinates.

Examples

julia> distance_2d(25, 42, 35, 80)
39.293765408777

julia> p1 = Point(25, 42)
Point(25, 42, nothing)

julia> p2 = Point(35, 80)
Point(35, 80, nothing)

julia> distance_2d(p1, p2)
39.293765408777

Polymath.Geometry.distance_3d — Function

distance_3d(x1, y1, z1, x2, y2, z2)
distance_3d(p1, p2)

Calculate distance between a pair of 3D points given the coordinates.

Examples

julia> distance_3d(14, 8, 5, 17, 38, 23)
35.11409973215888

julia> p1 = Point(14, 8, 5)
Point(14, 8, 5)

julia> p2 = Point(17, 38, 23)
Point(17, 38, 23)

julia> distance_3d(p1, p2)
35.11409973215888